* master:
  Modify uri_query pointer value to reference.
  tizenrt: Update Toolchain for next release
  Linux:ipadapter:send_msg: init control buffer
  oc_introspection update
  introspection
  oc_endpoint_t: Add local address, interface index
  Add few error checks to oc_core_res and oc_tls
  Make network up/down and session events optional
  Fix crash in make test
  Remove dependency in header files.

+++ api/oc_client_api.c
@@ -53,7 +53,7 @@ dispatch_coap_request(void)
         request_buffer, 0, (uint32_t)OC_BLOCK_SIZE, &block_size);
       if (payload) {
         coap_set_payload(request, payload, block_size);
-        coap_set_header_block1(request, 0, 1, block_size);
+        coap_set_header_block1(request, 0, 1, (uint16_t)block_size);
         coap_set_header_size1(request, payload_size);
         request->type = COAP_TYPE_CON;
         client_cb->qos = HIGH_QOS;
